requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
attrs==23.2.0
certifi==2024.2.2
click==8.1.7
distlib==0.3.8
filelock==3.13.1
frozenlist==1.4.1
idna==3.7
multidict==6.0.5
paho-mqtt==2.1.0
platformdirs==4.2.0
PyYAML==6.0.1
telepot==12.7
urllib3==2.2.1
virtualenv==20.25.1
yarl==1.9.4
